+++ lib/mdx.mli
@@ -17,7 +17,7 @@
 (** [Mdx] is a library to manipulate markdown code blocks.
 
     [mdx] allows to execute code blocks inside markdown files. The
-   supported code {{!Block}blocks} are either {{!Cram}cram-like}
+   supported code {{!module-Block}blocks} are either {{!module-Cram}cram-like}
    tests, raw OCaml fragments or {{!Toplevel}toplevel} phrases.
 
     Cram tests and toplevel phrases are sequences of commands and