+++ CHANGES.md
@@ -3,6 +3,9 @@
 #### Added
 
 - Report all parsing errors in Markdown files (#389, @NathanReb)
+- Add alternative syntax for explicitly setting the block-type.
+  The new label `block-type=...` can be set to `ocaml`, `toplevel`, `cram` or
+  `include`. (#385, @NathanReb)
 
 #### Changed